nlp/news-article-summarizer-sentiment-analysis/News-Article-Summarizer-with-Sentiment-Analysis/requirements.txt
streamlit
textblob
lxml
lxml_html_clean
newspaper3k
nltk
